+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -49,6 +49,12 @@ Notes](../../RELEASENOTES.md).
   boxing.
   ([#7645](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7645))
 
+* The OTLP exporter will now export logger version if it was specified. Log
+  records are grouped into `ScopeLogs` by logger name and version, so loggers
+  sharing a name but reporting different versions are exported as separate
+  scopes.
+  ([#7636](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7636))
+
 ## 1.17.0
 
 Released 2026-Jul-16